Functions for blackbody thermal emission (and its derivatives) calculation

TestingData — Module

This module contains some hardcoded data obtained from Berger, Chris G. 'Blackbody Calculator', Optics: The Website, https://www.opticsthewebsite.com/OpticsCalculators . Accessed on 5/14/2025

Main.TestingData.benchmark_data — Constant

The data for testing was obtained from

Berger, Chris G. 'Blackbody Calculator', Optics: The Website, https://www.opticsthewebsite.com/OpticsCalculators . Accessed on 5/14/2025

Main.TestingData.read_temperature_data — Method

read_temperature_data(tag)

Functions to read benchmark data spectrum from the default testing data folder

Main.TestingData.temperatures — Method

temperatures()

Returns numerical array of all temperatures in benchmark

Main.TestingData.testing_data — Method

testing_data(tag::String)

returns benchmark data for specified tag